ECHO-1804: added tps test

+++ scripts/node_deployer/deployer.py
@@ -71,6 +71,7 @@ class deployer:
         self.echo_bin = echo_bin
         self.echo_data_dir = "./tmp/echorand_test_datadir"
         self.image = image
+        self.pumba_started = False
 
         self.port = 13375
         self.rpc_port = 8090
@@ -204,6 +205,7 @@ class deployer:
         return self.node_names
 
     def run_pumba(self, nodes, time, jitter):
+        self.pumba_started = True
         cmd = "{pumba_bin} netem --interface=eth0 --duration 90m delay --time {time} --jitter {jitter} --correlation 0 {containers}"
         self.pumba_proc = subprocess.Popen(cmd.format(pumba_bin=self.pumba_bin,
             time=time, jitter=jitter, containers = nodes), shell=True, preexec_fn=os.setsid)
@@ -216,7 +218,8 @@ class deployer:
         print("Node deploying - Done")
 
     def kill_pumba(self):
-        os.killpg(os.getpgid(self.pumba_proc.pid), signal.SIGTERM)
+        if self.pumba_started == True:
+            os.killpg(os.getpgid(self.pumba_proc.pid), signal.SIGTERM)
 
 def test():
     d = deployer(node_count=2, echo_bin="/home/pplex/echo/build/bin/echo_node", pumba_bin="/home/pplex/pumba/.bin/pumba", image="ubuntu_delay")
